[3.12] gh-108303: Move more typing related files to Lib/test/typinganndata(GH-111825)
